+++ NosTests/Service/NIP19/NIP19EntityTests.swift
@@ -20,12 +20,29 @@ class NIP19EntityTests: XCTestCase {
         }
     }
 
-    // swiftlint:disable line_length
-    /// Example taken from [#1231](https://github.com/planetary-social/nos/issues/1231), which points to
-    /// [this note](https://njump.me/nevent1qqsqq0wah49rd6hjpezm275vys8pu6l5lcqddj9mz8cwrwf3m00k56gzyqalp33lewf5vdq847t6te0wvnags0gs0mu72kz8938tn24wlfze6a2cs3x)
+    /// Example taken from [#1234](https://github.com/planetary-social/nos/issues/1234) which points to
+    /// note1ar8gnwj6ugxyq6w9aqur74s2sn4dttk6zx0d25p34fsg476kqrksmkqqnr
+    @MainActor func test_nprofile_2() throws {
+        // swiftlint:disable:next line_length
+        let nprofile = "nprofile1qyvhwumn8ghj7un9d3shjtnndehhyapwwdhkx6tpdshszrnhwden5te0dehhxtnvdakz7qghwaehxw309aex2mrp0yhxummnw3ezucnpdejz7qg4waehxw309aex2mrp0yhxgctdw4eju6t09uq3xamnwvaz7tm0venxx6rpd9hzuur4vghszxnhwden5te0wfjkccte9eeks6t5vehhycm99ehkuef0qy08wumn8ghj7mn0wd68yttsw43zuam9d3kx7unyv4ezumn9wshszxrhwden5te0wfjkccte9e3h2unjv4h8gtnx095j7qgawaehxw309ahx7um5wghx6at5d9h8jampd3kx2apwvdhk6tcqyqe4dhnpkwty0ycuazepgzet4wphuzqscrh4zka7jt0qyjqypw9a60jrgzx"
+
+        let entity = try NIP19Entity.decode(bech32String: nprofile)
+        switch entity {
+        case .nprofile(let publicKey, let relays):
+            XCTAssertEqual(publicKey, "3356de61b39647931ce8b2140b2bab837e0810c0ef515bbe92de0248040b8bdd")
+            XCTAssertEqual(relays.count, 9)
+            let firstRelay = try XCTUnwrap(relays.first)
+            XCTAssertEqual(firstRelay, "wss://relay.snort.social/")
+        default:
+            XCTFail("Expected to get a nprofile")
+        }
+    }
+
+    /// Example taken from [#1231](https://github.com/planetary-social/nos/issues/1231) which points to
+    /// note1qq7am022xm40yrj9k4agcfqwre4lflsq6mytky0suxunrk7ldf5sjwth8x
     @MainActor func test_nevent() throws {
+        // swiftlint:disable:next line_length
         let nevent = "nevent1qyt8wumn8ghj7un9d3shjtnddaehgu3wwp6kytcpz9mhxue69uhkummnw3ezumrpdejz7qg4waehxw309aex2mrp0yhxgctdw4eju6t09uq3wamnwvaz7tmjv4kxz7fwwpexjmtpdshxuet59uq32amnwvaz7tmwdaehgu3wdau8gu3wv3jhvtcpr4mhxue69uhkummnw3ezucnfw33k76twv4ezuum0vd5kzmp0qyv8wumn8ghj7mn0wd68ytnxd46zuamf0ghxy6t69uq3jamnwvaz7tmjv4kxz7fwwdhx7un59eek7cmfv9kz7qghwaehxw309aex2mrp0yhxummnw3ezucnpdejz7qg3waehxw309ahx7um5wgh8w6twv5hsqg9p8569xea0fgnv0zuqnt3wsk5mu9j6xal7ten6332pg9r5h8g32gl7wn5w"
-        // swiftlint:enable line_length
 
         let entity = try NIP19Entity.decode(bech32String: nevent)
         switch entity {
@@ -68,7 +85,28 @@ class NIP19EntityTests: XCTestCase {
         }
     }
 
-    // From note12h0l6emckxgdfjnl3pfvyss4gp03yta7k0n9uwghywksksyfr9ws8q6war
+    /// Example taken from note16fddcyxfldwre2zywr96fnkmk7n3rtf4ehntdwy8ltyqgfntwfnq0dm347
+    @MainActor func test_nevent_3() throws {
+        // swiftlint:disable:next line_length
+        let nevent = "nevent1qy2hwumn8ghj7un9d3shjtnyv9kh2uewd9hj7qgewaehxw309aex2mrp0yh8xmn0wf6zuum0vd5kzmp0qyd8wumn8ghj7urewfsk66ty9enxjct5dfskvtnrdakj7qgwwaehxw309ahx7uewd3hkctcprdmhxue69uhkummnw3e8gctvdvhxummnw3erztnrdakj7qgkwaehxw309ajkgetw9ehx7um5wghxcctwvshszymhwden5te0wp6hyurvv4cxzeewv4ej7qghwaehxw309aex2mrp0yhxummnw3ezucnpdejz7qghwaehxw309aex2mrp0yh8qunfd4skctnwv46z7qgnwaehxw309aex2mrp0yhxvdm69e5k7tcqypr4d8t87evy6gka9xce8asxfr2kztrplttejhxzzamhkptceghxzxkjapf"
+
+        let entity = try NIP19Entity.decode(bech32String: nevent)
+        switch entity {
+        case .nevent(let eventID, let relays, let publicKey, let kind):
+            XCTAssertEqual(eventID, "47569d67f6584d22dd29b193f60648d5612c61fad7995cc217777b0578ca2e61")
+
+            XCTAssertEqual(relays.count, 10)
+            let firstRelay = try XCTUnwrap(relays.first)
+            XCTAssertEqual(firstRelay, "wss://relay.damus.io/")
+
+            XCTAssertNil(publicKey)
+            XCTAssertNil(kind)
+        default:
+            XCTFail("Expected to get a nevent")
+        }
+    }
+
+    /// Example taken from note12h0l6emckxgdfjnl3pfvyss4gp03yta7k0n9uwghywksksyfr9ws8q6war
     @MainActor func test_naddr() throws {
         // swiftlint:disable:next line_length
         let naddr = "naddr1qqjrsdf4xs6nvdrz95unsery956rswrx95unxvee94skvvp5xymkgwfcx9snyqg3waehxw309ahx7um5wgh8w6twv5hsygx0gknt5ymr44ldyyaq0rn3p5jpzkh8y8ymg773a06ytr4wldxz55psgqqqwense4rlem"
@@ -87,7 +125,7 @@ class NIP19EntityTests: XCTestCase {
         }
     }
 
-    // From note1xsems9u6xqfxl3hd3z4u4yr67vvf3g6w5l5vxwl8vqwcp0kgm2hsg3zf2w
+    /// Example taken from note1xsems9u6xqfxl3hd3z4u4yr67vvf3g6w5l5vxwl8vqwcp0kgm2hsg3zf2w
     @MainActor func test_naddr_2() throws {
         // swiftlint:disable:next line_length
         let naddr = "naddr1qvzqqqrujgpzp75cf0tahv5z7plpdeaws7ex52nmnwgtwfr2g3m37r844evqrr6jqyghwumn8ghj7vf5xqhxvdm69e5k7tcpzdmhxue69uhhqatjwpkx2urpvuhx2ue0qythwumn8ghj7un9d3shjtnswf5k6ctv9ehx2ap0qy2hwumn8ghj7un9d3shjtnyv9kh2uewd9hj7qg6waehxw309ac8junpd45kgtnxd9shg6npvchxxmmd9uq3xamnwvaz7tmjv4kxz7fwvcmh5tnfduhsz9thwden5te0wfjkccte9ejhs6t59ec82c30qyf8wumn8ghj7un9d3shjtn5dahkcue0qy88wumn8ghj7mn0wvhxcmmv9uqpqvenx56njvpnxqcrsdf4xqcrwdqufrevn"