+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Use NIP-05 for shared links to profile.
+- Decode nprofile, nevent, and naddr NIP-19 entities.
 
 ## [0.1.19] - 2024-07-01Z